# backend/domain/schemas/checkin.py
"""
backend/domain/schemas/checkin.py
"""

from __future__ import annotations

from datetime import date, datetime
from typing import Any, Literal, Optional
from uuid import UUID

from pydantic import BaseModel, ConfigDict, Field


class CheckinMessage(BaseModel):
    model_config = ConfigDict(extra="forbid")

    role: Literal["user", "assistant"]
    content: str


class DailyTaskOut(BaseModel):
    model_config = ConfigDict(extra="forbid")

    id: UUID
    title: str
    description: Optional[str] = None
    is_completed: bool
    completed_at: Optional[datetime] = None
    status: Literal["active", "suspended"] = "active"
    due_date: Optional[date] = None
    created_at: Optional[datetime] = None
    checkin_date: Optional[date] = None
    checkin_session_id: Optional[UUID] = None


# Backwards-compatible alias
WeeklyTaskOut = DailyTaskOut


class CheckinSessionOut(BaseModel):
    model_config = ConfigDict(extra="forbid")

    id: UUID
    checkin_date: date
    status: str
    summary: Optional[str] = None
    mood_score: Optional[int] = None
    energy_level: Optional[int] = None
    motivation_level: Optional[int] = None
    workload_level: Optional[str] = None
    main_blocker: Optional[str] = None
    stage: Optional[str] = "opening"
    turn_count: int = 0
    messages: list[dict[str, Any]] = Field(default_factory=list)
    daily_tasks: list[DailyTaskOut] = Field(default_factory=list)


class TaskCompleteRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    is_completed: bool = True


class MoodUpdateRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    mood_score: int = Field(ge=1, le=5)


class CheckinHistoryItem(BaseModel):
    model_config = ConfigDict(extra="forbid")

    id: UUID
    checkin_date: date
    status: str
    summary: Optional[str] = None
    mood_score: Optional[int] = None
    task_count: int = 0
    completed_task_count: int = 0


class ProfileStatsOut(BaseModel):
    model_config = ConfigDict(extra="forbid")

    total_checkins: int
    streak_days: int
    completed_tasks: int
    open_tasks: int = 0
    capacity_history: list[dict[str, Any]] = Field(default_factory=list)