# specs/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Equa API
  version: 1.0.0
  description: |
    Equa B2B2C MVP API contract.
    Breaking changes require a new major path (`/api/v2`).
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/checkins:
    post:
      tags: [checkins]
      summary: Start or resume the current daily check-in
      operationId: startCheckin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current check-in session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinSessionResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/checkins/current:
    get:
      tags: [checkins]
      summary: Get or start the current daily check-in
      operationId: getCurrentCheckin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current check-in session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinSessionResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/checkins/history:
    get:
      tags: [checkins]
      summary: List past daily check-in sessions
      operationId: listCheckinHistory
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        '200':
          description: Check-in history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinHistoryResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/checkins/current/mood:
    patch:
      tags: [checkins]
      summary: Set mood score for the current check-in (1-5)
      operationId: updateCurrentMood
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoodUpdateRequest'
      responses:
        '200':
          description: Updated check-in session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinSessionResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/tasks:
    get:
      tags: [checkins]
      summary: List all daily tasks for the current student
      operationId: listTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/tasks/{task_id}:
    patch:
      tags: [checkins]
      summary: Mark a weekly task complete/incomplete
      operationId: patchTask
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPatchResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/institution/students:
    get:
      tags: [institution]
      summary: List students with risk scores (staff only)
      operationId: listInstitutionStudents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student risk rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionStudentsResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/roi:
    get:
      tags: [institution]
      summary: ROI metrics for the tenant (staff only)
      operationId: getInstitutionRoi
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ROI metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoiResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/overview:
    get:
      tags: [institution]
      summary: Tenant overview metrics (staff only)
      operationId: getInstitutionOverview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Overview metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/usage:
    get:
      tags: [institution]
      summary: Adoption / usage metrics (staff only)
      operationId: getInstitutionUsage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/me:
    get:
      tags: [institution]
      summary: Staff profile + tenant info
      operationId: getInstitutionMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Staff profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/settings:
    patch:
      tags: [institution]
      summary: Update tenant settings (admin only)
      operationId: patchInstitutionSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revenue_per_student]
              properties:
                revenue_per_student:
                  type: number
                  exclusiveMinimum: 0
                  maximum: 1000000
      responses:
        '200':
          description: Updated tenant settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/students/{student_id}:
    get:
      tags: [institution]
      summary: Student XAI detail — metrics only, never raw chat
      operationId: getInstitutionStudent
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Student risk detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/institution/assistant/stream:
    post:
      tags: [institution]
      summary: Stream institution AI assistant (metrics-only context)
      operationId: streamInstitutionAssistant
      description: |
        Staff-only SSE assistant. Context pack contains aggregate metrics and
        XAI rationales only — never raw student chat messages.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: array
          items: {}
          default: []
    SuccessEnvelope:
      type: object
      additionalProperties: false
      required: [success, data, error, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
        error:
          nullable: true
          type: object
        meta:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, data, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          type: object
          additionalProperties: true
...
    DailyTaskOut:
      type: object
      additionalProperties: false
      required: [id, title, is_completed]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          nullable: true
          type: string
        is_completed:
          type: boolean
        completed_at:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
          enum: [active, suspended]
          default: active
        due_date:
          nullable: true
          type: string
          format: date
        created_at:
          nullable: true
          type: string
          format: date-time
        checkin_date:
          nullable: true
          type: string
          format: date
        checkin_session_id:
          nullable: true
          type: string
          format: uuid
    CheckinSessionOut:
      type: object
      additionalProperties: false
      required: [id, checkin_date, status, messages, daily_tasks]
      properties:
        id:
          type: string
          format: uuid
        checkin_date:
          type: string
          format: date
        status:
          type: string
        summary:
          nullable: true
          type: string
        mood_score:
          nullable: true
          type: integer
          minimum: 1
          maximum: 5
        messages:
          type: array
          items:
            type: object
        daily_tasks:
          type: array
          items:
            $ref: '#/components/schemas/DailyTaskOut'
    CheckinSessionResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CheckinSessionOut'
    MoodUpdateRequest:
      type: object
      additionalProperties: false
      required: [mood_score]
      properties:
        mood_score:
          type: integer
          minimum: 1
          maximum: 5
    CheckinHistoryItem:
      type: object
      additionalProperties: false
      required: [id, checkin_date, status, task_count, completed_task_count]
      properties:
        id:
          type: string
          format: uuid
        checkin_date:
          type: string
          format: date
        status:
          type: string
        summary:
          nullable: true
          type: string
        mood_score:
          nullable: true
          type: integer
        task_count:
          type: integer
        completed_task_count:
          type: integer
    CheckinHistoryResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                sessions:
                  type: array
                  items:
                    $ref: '#/components/schemas/CheckinHistoryItem'
    TaskListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                tasks:
                  type: array
                  items:
                    $ref: '#/components/schemas/DailyTaskOut'
    TaskCompleteRequest:
      type: object
      additionalProperties: false
      properties:
        is_completed:
          type: boolean
          default: true
    TaskPatchResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              additionalProperties: true
...
    StudentRiskRow:
      type: object
      additionalProperties: false
      required: [user_id, email, risk_level]
      properties:
        user_id:
          type: string
          format: uuid
        full_name:
          nullable: true
          type: string
        email:
          type: string
        risk_level:
          type: string
          enum: [green, yellow, red]
        rationale:
          nullable: true
          type: string
        metrics:
          nullable: true
          type: object
        capacity_score:
          nullable: true
          type: number
        updated_at:
          nullable: true
          type: string
          format: date-time
    InstitutionStudentsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [students]
              properties:
                students:
                  type: array
                  items:
                    $ref: '#/components/schemas/StudentRiskRow'
    RoiMetrics:
      type: object
      additionalProperties: false
      required:
        - prevented_dropouts
        - protected_revenue
        - revenue_per_student
        - active_high_risk
        - total_students
      properties:
        prevented_dropouts:
          type: integer
        protected_revenue:
          type: number
        revenue_per_student:
          type: number
        active_high_risk:
          type: integer
        total_students:
          type: integer
    RoiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RoiMetrics'